Large overhaul to task manager design, now jobs can optionally be run batched!

+++ task_manager.py
@@ -41,55 +41,107 @@ def check_directory():
         os.remove(config['loc'])
         sys.exit()
 
-    files = [file for file in files if config['key_name'] in file]
-    
-    valid = [str(i) for i in range(config['start_key_num'], config['start_key_num']+config['num_jobs'])]
-    valid_files = []
-    
-    for file in files:
-        name = file.split('.')[0].replace(config['key_name'], '')
-        
-        if name in valid:
-            valid_files.append(file)        
+    if config['single_jobs']:
 
-    return valid_files
+        valid_files = []
+        files = [file for file in files if config['key_name'] in file]
+        valid = [str(i) for i in range(config['start_key_num'], config['start_key_num']+config['num_jobs'])]
+    
+        for file in files:
+            name = file.split('.')[0].replace(config['key_name'], '')
+            
+            if name in valid:
+                valid_files.append(file)        
+
+        return valid_files
+
+    else:
+        
+        files = [file for file in files if config['key_name'] + '_Output' in file]
+        return files
+
+def read_file(loc):
+
+    with open(loc, 'r') as f:
+        lines = f.readlines()
+    return lines
+
+def write_file(loc, lines):
+
+    with open(loc, 'w') as f:
+        for line in lines:
+            f.write(line)
+
+def change_lines(lines, replacements):
+
+    for i in range(len(lines)):
+        for r in replacements:
+            if r in lines[i]:
+                lines[i].replace(r, replacements[r])
+    
+    return lines
 
 def change_temp_script(name, load):
     
-    with open(config['template_name'], 'r') as f:
-        lines = f.readlines()
-        
-    for i in range(len(lines)):
-        
-        if 'MAX_RUN_TIME' in lines[i]:
-            lines[i] = lines[i].replace('MAX_RUN_TIME', str(config['max_run_time']))
-        if 'EV_SEARCH_LOCATION' in lines[i]:
-            lines[i] = lines[i].replace('EV_SEARCH_LOCATION', config['ev_search_dr'])
-        if 'REPLACE' in lines[i]:
-            pkl_loc = os.path.join(config['key_dr'], config['key_name'] + name)
-            lines[i] = lines[i].replace('REPLACE', pkl_loc)
-        if 'LOAD' in lines[i]:
-            lines[i] = lines[i].replace('LOAD', load)
-        if 'CONFIG_LOC' in lines[i]:
-            lines[i] = lines[i].replace('CONFIG_LOC', CONFIG_LOC)
-            
-            
-    with open(config['temp_name'], 'w') as f:
-        for line in lines:
-            f.write(line)
-            
-def run_job():
-    os.system('qsub ' + config['temp_name'])
+    pkl_loc = os.path.join(config['key_dr'], config['key_name'] + name)
+
+    replacements = {'MAX_RUN_TIME': str(config['max_run_time'],
+                    'EV_SEARCH_LOCATION': config['ev_search_dr'],
+                    'REPLACE': pkl_loc,
+                    'NAME': config['key_name'] + name,
+                    'LOAD': load,
+                    'CONFIG_LOC': CONFIG_LOC}
+
+    lines = read_file(config['template_name'])
+    lines = change_lines(lines, replacements)
+    write_file(config['temp_name'], lines)
+
+def change_temp_scripts(load):
+
+    replacements = {'MAX_RUN_TIME': str(config['max_run_time'],
+                    'EV_SEARCH_LOCATION': config['ev_search_dr'],
+                    'NAME': config['key_name'] + '_Output',
+                    'SUBJOB_NAME': config['subjob_name'],
+                    'NUM_JOBS': config['num_jobs'],
+                    'START_NUM': config['start_num'],
+                    'LOAD': load,
+                    'CONFIG_LOC': CONFIG_LOC}
+
+    #Master
+    lines = read_file(config['master_template'])
+    lines = change_lines(lines, replacements)
+    write_file(config['master_name'], lines)
+
+    #Subjob
+    lines = read_file(config['subjob_template'])
+    lines = change_lines(lines, replacements)
+    write_file(config['subjob_name'], lines)
+    
+    #Change permissions of subjob
+    os.system('chmod u+x ' + config['subjob_name'])
+
+def run_job(temp_name):
+    os.system('qsub ' + temp_name)
 
 def init_jobs(arg):
+
+    if config['single_jobs']:
     
-    for i in range(config['start_key_num'], config['start_key_num']+config['num_jobs']):
-        change_temp_script(str(i), arg)
-        run_job()
-        
-        COUNTER[str(i)] = 1
-        TIMER[str(i)] = time.time()
-        
+        for i in range(config['start_key_num'], config['start_key_num']+config['num_jobs']):
+            change_temp_script(str(i), arg)
+            run_job(config['temp_name'])
+            
+            COUNTER[str(i)] = 1
+            TIMER[str(i)] = time.time()
+
+    else:
+
+        change_temp_scripts(arg)
+        run_job(config['master_name'])
+
+        COUNTER['1'] = 1
+        TIMER['1'] = time.time()
+
 def check_counter():
     '''Returns true if jobs need to be run more'''
     
@@ -101,18 +153,31 @@ def check_counter():
     return False
 
 def proc_new_files(files):
+
+    if config['single_jobs']:
     
-    for file in files:
-        name = file.split('.')[0].replace(config['key_name'], '')
-        
-        if COUNTER[name] < config['num_search_gens']:
-            change_temp_script(name, '1')
-            run_job()
-        
-            COUNTER[name] += 1
-            TIMER[name] = time.time()
+        for file in files:
+            name = file.split('.')[0].replace(config['key_name'], '')
             
-            os.remove(file)
+            if COUNTER[name] < config['num_search_gens']:
+                change_temp_script(name, '1')
+                run_job(config['temp_name'])
+            
+                COUNTER[name] += 1
+                TIMER[name] = time.time()
+                
+                os.remove(file)
+
+    else:
+        
+        file = files[0]
+        if COUNTER['1'] < config['num_search_gens']:
+            change_temp_scripts('1')
+            run_job(config['master_name'])
+
+            COUNTER['1'] += 1
+            TIMER['1'] = time.time()
+
         
 def save_progress():
     with open(config['progress_name'], 'w') as f:
@@ -122,37 +187,62 @@ def save_progress():
 def restart_job(name):
     
     #If starting from scratch and counter still == 1
-    if config['preloaded'] == False and COUNTER[name] == 1 and config['continue'] == False:
+    if config['preloaded'] == False and COUNTER[name] == 1:
         print('restart job ', name, 'with new')
 
         change_temp_script(name, '0')
-        run_job()
-        
         COUNTER[name] = 1
-        TIMER[name] = time.time()
     
     #Otherwise, just try again
     else:
         print('restart job w/ load', name)
+        
         change_temp_script(name, '1')
-        run_job()
-    
         COUNTER[name] += 1
-        TIMER[name] = time.time()
+    
+    run_job(config['temp_name'])
+    TIMER[name] = time.time()
+
+def restart_jobs():
+
+    #If starting from scratch and counter still == 1
+    if config['preloaded'] == False and COUNTER['1'] == 1:
+        print('restart jobs with new')
+
+        change_temp_scripts('0')
+        COUNTER['1'] = 1
+        
+    #Otherwise, just try again
+    else:
+        print('restart jobs with load')
+
+        change_temp_scripts('1')
+        COUNTER['1'] += 1
+
+    run_job(config['master_name'])
+    TIMER['1'] = time.time()
             
 def check_progress():
     
     current_time = time.time()
-    
-    for i in range(config['start_key_num'], config['start_key_num']+config['num_jobs']):
-        i = str(i)
-        
-        if COUNTER[i] < config['num_search_gens']:
-            print('check progress job ', i, current_time - TIMER[i])
+
+    if config['single_jobs']:
+        for i in range(config['start_key_num'], config['start_key_num']+config['num_jobs']):
+            i = str(i)
             
-            #If job hasnt been updated in enough time, restart it-
-            if current_time - TIMER[i] > config['restart_lim'] * 3600:
-                restart_job(i)
+            if COUNTER[i] < config['num_search_gens']:
+                print('check progress job ', i, current_time - TIMER[i])
+                
+                #If job hasnt been updated in enough time, restart it-
+                if current_time - TIMER[i] > config['restart_lim'] * 3600:
+                    restart_job(i)
+
+    else:
+        if COUNTER['1'] < config['num_search_gens']:
+            print('check progress on batched jobs ', current_time - TIMER['1'])
+
+            if current_time - TIMER['1'] > config['restart_lim'] * 3600:
+                restart_jobs()
 
 #Ensure key directory exists
 key_dr = os.path.join(config['ev_search_dr'], config['key_dr'])
@@ -164,20 +254,11 @@ if not os.path.isfile(config['loc']) or config['create_new_data']:
     print('creating data file')    
     process_new_dataset(config)
 
-if config['continue']:
-    for i in range(config['start_key_num'], config['start_key_num']+config['num_jobs']):
-        COUNTER[str(i)] = 0
-        TIMER[str(i)] = time.time()
-
-elif config['preloaded']:
-    init_jobs('1')
-
-else:
-    init_jobs('0')
+init_jobs(str(int(config['preloaded'])))
 
 sys.stdout.flush()
 save_progress()
-  
+
 #MAIN LOOP
 while check_counter():
     
@@ -191,4 +272,9 @@ while check_counter():
     check_progress()
         
     time.sleep(config['check_every'])
-    sys.stdout.flush()        
+    sys.stdout.flush()
+
+
+
+
+
